Add files via upload
This MS Excel Project provides us with complete insights on a specific social media content dataset . It answers several questions that may be asked by the client. The bar and pie  charts helps us visualize the complex data in a simple manner which can be understood by all the members involved in the project.